@@ -6,7 +6,7 @@
         "Icinga",
         "Nagios"
     ],
-    "homepage": "https://www.random-host.com",
+    "homepage": "https://github.random-host.tv",
     "license": "BSD-3-Clause",
     "authors": [
         {
@@ -16,20 +16,26 @@
     ],
     "config": {
         "platform": {
-            "php": "5.3.3"
+            "php": "7.3.31"
         }
     },
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=7.3.0"
     },
     "require-dev": {
-        "behat/behat": "~2.0",
-        "phpunit/phpunit": "~4.0",
-        "phpspec/phpspec": "~2.4"
+        "phpunit/phpunit": "~9.0"
     },
     "autoload": {
         "psr-4": {
             "randomhost\\Icinga\\": "src/php/"
         }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "randomhost\\Icinga\\Tests\\": "src/tests/php"
+        }
+    },
+    "scripts": {
+        "test": "vendor/bin/phpunit"
     }
 }
