@@ -1,16 +1,20 @@
 <?php
+
+declare(strict_types=1);
+
 namespace randomhost\Icinga\Check;
 
-use Exception;
 use randomhost\Icinga\Base as IcingaBase;
+use randomhost\Icinga\Plugin;
 
 /**
  * Base class for Icinga check plugins.
  *
  * @author    Ch'Ih-Yu <chi-yu@web.de>
- * @copyright 2016 random-host.com
- * @license   http://www.debian.org/misc/bsd.license BSD License (3 Clause)
- * @link      http://github.random-host.com/icinga/
+ * @copyright 2022 Random-Host.tv
+ * @license   https://opensource.org/licenses/BSD-3-Clause BSD License (3 Clause)
+ *
+ * @see https://github.random-host.tv
  */
 abstract class Base extends IcingaBase implements Check
 {
@@ -19,16 +23,18 @@ abstract class Base extends IcingaBase implements Check
      *
      * @return $this
      */
-    public function run()
+    public function run(): Plugin
     {
         try {
             return $this
                 ->preRun()
-                ->check();
-        } catch (Exception $e) {
+                ->check()
+            ;
+        } catch (\Exception $e) {
             return $this
                 ->setMessage($e->getMessage())
-                ->setCode($e->getCode());
+                ->setCode($e->getCode())
+            ;
         }
     }
 
@@ -37,5 +43,5 @@ abstract class Base extends IcingaBase implements Check
      *
      * @return $this
      */
-    abstract protected function check();
+    abstract protected function check(): Plugin;
 }
