@@ -1,34 +1,53 @@
 <?php
+
+declare(strict_types=1);
+
 namespace randomhost\Icinga\Notification;
 
-use Exception;
 use randomhost\Icinga\Base as IcingaBase;
+use randomhost\Icinga\Plugin;
 
 /**
  * Base class for Icinga notification plugins.
  *
  * @author    Ch'Ih-Yu <chi-yu@web.de>
- * @copyright 2016 random-host.com
- * @license   http://www.debian.org/misc/bsd.license BSD License (3 Clause)
- * @link      http://github.random-host.com/icinga/
+ * @copyright 2022 Random-Host.tv
+ * @license   https://opensource.org/licenses/BSD-3-Clause BSD License (3 Clause)
+ *
+ * @see https://github.random-host.tv
  */
 abstract class Base extends IcingaBase implements Notification
 {
+    /**
+     * Command line options available to be used in messages.
+     */
+    protected const MESSAGE_OPTIONS = [
+        'type',
+        'service',
+        'host',
+        'address',
+        'state',
+        'time',
+        'output',
+    ];
+
     /**
      * Sends the Icinga notification.
      *
      * @return $this
      */
-    public function run()
+    public function run(): Plugin
     {
         try {
             return $this
                 ->preRun()
-                ->send();
-        } catch (Exception $e) {
+                ->send()
+            ;
+        } catch (\Exception $e) {
             return $this
                 ->setMessage($e->getMessage())
-                ->setCode($e->getCode());
+                ->setCode($e->getCode())
+            ;
         }
     }
 
@@ -37,5 +56,5 @@ abstract class Base extends IcingaBase implements Notification
      *
      * @return $this
      */
-    abstract protected function send();
+    abstract protected function send(): Plugin;
 }
