@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace randomhost\Icinga\Notification;
 
 use randomhost\Icinga\Plugin;
@@ -7,9 +10,10 @@ use randomhost\Icinga\Plugin;
  * Interface definition for Icinga notification plugins.
  *
  * @author    Ch'Ih-Yu <chi-yu@web.de>
- * @copyright 2016 random-host.com
- * @license   http://www.debian.org/misc/bsd.license BSD License (3 Clause)
- * @link      http://github.random-host.com/icinga/
+ * @copyright 2022 Random-Host.tv
+ * @license   https://opensource.org/licenses/BSD-3-Clause BSD License (3 Clause)
+ *
+ * @see https://github.random-host.tv
  */
 interface Notification extends Plugin
 {
